// packages/plugin-clientportal-ui/src/types.ts
export type OTPConfig = {
  emailSubject?: string;
  content: string;
  smsTransporterType?: '' | 'messagePro';
  codeLength: number;
  loginWithOTP: boolean;
  expireAfter: number;
};

export type MailConfig = {
  subject: string;
  registrationContent: string;
  invitationContent: string;
};

export type ManualVerificationConfig = {
  userIds: string[];
  verifyCustomer: boolean;
  verifyCompany: boolean;
};

export type PasswordVerificationConfig = {
  verifyByOTP: boolean;
  emailSubject: string;
  emailContent: string;
  smsContent: string;
};

export interface IVerificationRequest {
  status: string;
  attachments: IAttachment[];
  description?: string;
  verifiedBy?: string;
}

export interface IClientPortalUserDoc {
  firstName: string;
  companyName: string;
  companyRegistrationNumber: string;
  lastName: string;
  code: string;
  phone: string;
  email: string;
  username: string;
  ownerId?: string;
  saashqCustomerId: string;
  saashqCompanyId: string;
  clientPortalId: string;
  type: string;
  clientPortal: ClientPortalConfig;

  customer: ICustomer;
  company: ICompany;

  isPhoneVerified: boolean;
  isEmailVerified: boolean;

  lastSeenAt: Date;
  sessionCount: number;
  isOnline: boolean;

  customFieldsData: JSON;
  avatar: string;

  verificationRequest: IVerificationRequest;
}

export interface IClientPortalUser extends IClientPortalUserDoc {
  _id: string;
  createdAt: Date;
  modifiedAt: Date;
  forumSubscriptionEndsAfter?: string;
  clientPortal: ClientPortalConfig;
}
// ...
export type ClientPortalUserAssignCompanyMutationResponse = {
  clientPortalUserAssignCompany: (mutation: {
    variables: {
      userId: string;
      saashqCompanyId: string;
      saashqCustomerId: string;
    };
  }) => Promise<any>;
};

export type ClientPortalVerifyUsersMutationResponse = {
  clientPortalUsersVerify: (mutation: {
    variables: { type: string; userIds: string[] };
  }) => Promise<any>;
};

export type SocialpayConfig = {
  publicKey: string;
  certId: string;
};

export type ClientPortalConfig = {
  _id?: string;
  name?: string;
  description?: string;
  url?: string;
  logo?: string;
  icon?: string;
  domain?: string;
  dnsStatus?: string;
  messengerBrandCode?: string;
  knowledgeBaseLabel?: string;
  knowledgeBaseTopicId?: string;
  ticketLabel?: string;
  dealLabel?: string;
  purchaseLabel?: string;
  taskPublicBoardId?: string;
  taskPublicPipelineId?: string;
  taskPublicLabel?: string;
  taskLabel?: string;
  taskStageId?: string;
  taskBoardId?: string;
  taskPipelineId?: string;
  ticketStageId?: string;
  ticketBoardId?: string;
  ticketPipelineId?: string;
  dealStageId?: string;
  dealBoardId?: string;
  dealPipelineId?: string;
  purchaseStageId?: string;
  purchaseBoardId?: string;
  purchasePipelineId?: string;
  styles?: Styles;
  mobileResponsive?: boolean;
  googleCredentials?: object;
  googleClientId?: string;
  googleRedirectUri?: string;
  googleClientSecret?: string;
  facebookAppId?: string;
  saashqAppToken?: string;

  kbToggle?: boolean;
  publicTaskToggle?: boolean;
  ticketToggle?: boolean;
  dealToggle?: boolean;
  purchaseToggle?: boolean;
  taskToggle?: boolean;
  otpConfig?: OTPConfig;
  mailConfig?: MailConfig;
  manualVerificationConfig?: ManualVerificationConfig;
  passwordVerificationConfig?: PasswordVerificationConfig;

  testUserEmail: string;
  testUserPhone: string;
  testUserPassword: string;
  testUserOTP: number;

  tokenExpiration?: number;
  refreshTokenExpiration?: number;
  tokenPassMethod: 'cookie' | 'header';
  vendorParentProductCategoryId?: string;
  socialpayConfig?: SocialpayConfig;
};

export type Styles = {
  bodyColor?: string;
  headerColor?: string;
  footerColor?: string;
  helpColor?: string;
  backgroundColor?: string;
  activeTabColor?: string;
  baseColor?: string;
  headingColor?: string;
  linkColor?: string;
  linkHoverColor?: string;
  primaryBtnColor?: string;
  secondaryBtnColor?: string;
  dividerColor?: string;
  baseFont?: string;
  headingFont?: string;
  __typename?: string;
};